refactor: 상품 추가 및 상품 수정 코드 리팩토링
